This commit contains a comprehensive set of fixes and improvements based on a broad "debug" request. The changes address security vulnerabilities, code inconsistencies, and bugs, and improve the overall quality and robustness of the application.
The main changes are:
- Fixed an issue where the application was stuck on a preloader screen by providing a default language in the LanguageContext.
- Removed a hardcoded admin login backdoor and updated the admin login page to use the standard, more secure login flow.
- Replaced a hardcoded temporary password in the "create patron" feature with a randomly generated one.
- Added authorization checks to all admin pages to prevent unauthorized access.
- Refactored the codebase to use `role` instead of the inconsistent `user_type` for user roles. This change was applied to the database schema, API routes, and frontend components.
- Improved the security of the authentication system by adding robust validation for the JWT payload and ensuring the JWT_SECRET environment variable is always present.
