+++ .env.local
@@ -7,4 +7,4 @@ KV_REST_API_TOKEN="placeholder_token"
 MOCK_API=true
 
 # When 'true', this forces the language selection preloader to always be visible for testing.
-NEXT_PUBLIC_FORCE_SHOW_PRELOADER=true
+NEXT_PUBLIC_FORCE_SHOW_PRELOADER=false
