--- app/api/admin/login/route.ts
@@ -1,43 +1,0 @@
-import { NextRequest, NextResponse } from 'next/server';
-import { SignJWT } from 'jose';
-import { cookies } from 'next/headers';
-
-const JWT_SECRET = new TextEncoder().encode(process.env.JWT_SECRET!);
-const COOKIE_NAME = 'session';
-
-export async function POST(request: NextRequest) {
-  try {
-    const { username, password } = await request.json();
-
-    if (username === 'admin' && password === 'admin') {
-      const adminUser = {
-        id: 'admin_user',
-        role: 'admin' as const,
-        displayName: 'Admin',
-        avatar: 'https://i.pravatar.cc/150?u=admin',
-        sessionVersion: 1,
-      };
-
-      const token = await new SignJWT({ user: adminUser })
-        .setProtectedHeader({ alg: 'HS256' })
-        .setIssuedAt()
-        .setExpirationTime('24h')
-        .sign(JWT_SECRET);
-
-      cookies().set(COOKIE_NAME, token, {
-        httpOnly: true,
-        secure: process.env.NODE_ENV === 'production',
-        sameSite: 'lax',
-        path: '/',
-        maxAge: 60 * 60 * 24, // 1 day
-      });
-
-      return NextResponse.json({ success: true, message: 'Login successful' });
-    }
-
-    return NextResponse.json({ success: false, message: 'Invalid credentials' }, { status: 401 });
-  } catch (error) {
-    console.error('Admin login error:', error);
-    return NextResponse.json({ success: false, message: 'An internal server error occurred' }, { status: 500 });
-  }
-}
