+++ app/api/notifications/send/route.ts
@@ -13,7 +13,7 @@ if (process.env.VAPID_SUBJECT && process.env.NEXT_PUBLIC_VAPID_PUBLIC_KEY && pro
 
 export async function POST(request: NextRequest) {
   const payload = await verifySession();
-  if (!payload || !payload.user || payload.user.user_type !== 'admin') {
+  if (!payload || !payload.user || payload.user.role !== 'admin') {
     return NextResponse.json({ success: false, message: 'Unauthorized' }, { status: 403 });
   }
