+++ start.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+nohup yarn dev > app.log 2>&1 &
